LLaVA_flowertune/requirements.txt
flwr-nightly[rest,simulation]
flwr_datasets==0.0.2
hydra-core==1.3.2
scipy==1.11.2
fschat[model_worker,webui]==0.2.35
transformers>=4.39.0
peft
bitsandbytes
trl>=0.8.3
